apps/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
redis==5.2.0
python-multipart==0.0.12
mistralai
fal-client
requests
pydantic
python-decouple
